ready to pull

+++ Day6/Activity2/GrossSalary.py
@@ -1,4 +1,4 @@
 rate = 500
 
 def computeGrossSalary(hour):
-    return float(hour * rate)
+    return hour * rate
